Revert "vmalert: init unit test (#4596)"
This reverts commit da60a68d
