+++ app/vmalert/Makefile
@@ -74,7 +74,6 @@ test-vmalert:
 	go test -v -race -cover ./app/vmalert/config
 	go test -v -race -cover ./app/vmalert/remotewrite
 	go test -v -race -cover ./app/vmalert/utils
-	go test -v -race -cover ./app/vmalert/unittest
 
 run-vmalert: vmalert
 	./bin/vmalert -rule=app/vmalert/config/testdata/rules/rules2-good.rules \
@@ -103,10 +102,6 @@ replay-vmalert: vmalert
 		-external.label=replica=a \
 		-replay.timeFrom=2021-05-11T07:21:43Z \
 		-replay.timeTo=2021-05-29T18:40:43Z
-
-unittest-vmalert: vmalert
-	./bin/vmalert -unittestFile=app/vmalert/unittest/testdata/test1.yaml \
- 		-unittestFile=app/vmalert/unittest/testdata/test2.yaml
 
 vmalert-linux-amd64:
 	APP_NAME=vmalert CGO_ENABLED=1 GOOS=linux GOARCH=amd64 $(MAKE) app-local-goos-goarch
